requirements.txt
streamlit
pandas
plotly
office365-rest-python-client
openpyxl
